# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "open-api doc",
    "version": "v1"
  },
  "paths": {
    "/user-emails/{email}/verification-code": {
      "post": {
        "operationId": "requestEmailVerificationCode",
        "tags": [
          "PortalUserIdentifierVerificationController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDtoOfVoid"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "registerUser",
        "tags": [
          "PortalUserRegistrationController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDtoOfAccessTokenApiResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationApiRequest"
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/oauth2/token": {
      "post": {
        "operationId": "getAccessToken",
        "tags": [
          "AccessTokenController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDtoOfAccessTokenApiResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessTokenRequest"
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/password-resets/{email}": {
      "post": {
        "operationId": "requestPasswordResetWithEmail",
        "tags": [
          "PasswordResetRequestController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDtoOfVoid"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/password/{identifier}/{resetToken}": {
      "post": {
        "operationId": "resetPasswordWithResetToken",
        "tags": [
          "PasswordResetController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDtoOfAccessTokenApiResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetApiRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "identifier",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resetToken",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resetToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/key/{kid}": {
      "get": {
        "operationId": "getJsonWebKey",
        "tags": [
          "SignatureKeyController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDtoOfJwtWebTokenResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "kid",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/login": {
      "post": {
        "operationId": "login",
        "tags": [
          "LoginController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDtoOfAccessTokenApiResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/health": {
      "get": {
        "operationId": "health",
        "tags": [
          "IndexController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "nullable": true
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/api-docs": {
      "get": {
        "operationId": "apiDocs",
        "tags": [
          "IndexController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/": {
      "get": {
        "operationId": "index",
        "tags": [
          "IndexController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/clients": {
      "post": {
        "operationId": "createClient",
        "tags": [
          "ClientController"
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDtoOfClientResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientRegistrationRequest"
              }
            }
          }
        },
        "parameters": []
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponseDtoOfVoid": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "number"
          }
        }
      },
      "ApiResponseDtoOfAccessTokenApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AccessTokenApiResponse"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "number"
          }
        }
      },
      "AccessTokenApiResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "displayName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ]
          },
          "emailAddresses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": false
          },
          "phoneNumbers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": false
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataResponse"
            },
            "uniqueItems": false
          },
          "passwordUpdateRequired": {
            "type": "boolean"
          },
          "refresh_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          },
          "access_token": {
            "type": "string"
          },
          "secondsTillExpiry": {
            "type": "number"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DataResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "UserRegistrationApiRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "otherNames": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "phoneNumberVerificationCode": {
            "type": "string"
          },
          "emailVerificationCode": {
            "type": "string"
          },
          "isPasswordUpdateRequired": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDataApiRequest"
            },
            "uniqueItems": false
          }
        }
      },
      "UserDataApiRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "AccessTokenRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
      "PasswordResetApiRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "invalidateOtherSession": {
            "type": "boolean"
          }
        }
      },
      "ApiResponseDtoOfJwtWebTokenResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JwtWebTokenResponse"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "number"
          }
        }
      },
      "JwtWebTokenResponse": {
        "type": "object",
        "properties": {
          "kid": {
            "type": "string"
          },
          "use": {
            "type": "string"
          },
          "kty": {
            "type": "string"
          },
          "modulus": {
            "type": "string"
          },
          "exponent": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ApiResponseDtoOfClientResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ClientResponse"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "number"
          }
        }
      },
      "ClientResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        }
      },
      "ClientRegistrationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      }
    }
  }
}